requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
beautifulsoup4==4.12.2
yfinance==0.2.33
python-dotenv==1.0.0
pykrx>=1.0.45
